Update beebs parsing at long last
Had been putting that off for quite some time. Sucked it up. Hacky steps
to keep in mind

- Bunch of " are escaped in the source, I just dump all \ and hope for
  the best. Seems to work for now
- One of the property keys in the resulting json comes with a bunch of
  random garbage after it. I added a custom deserializer for that one.

Rest of it is pretty standard stuff compared to the old version, I
think. Not sure I quite like the naming of leagues yet (combo of
tournament/stage/round names), but I might be inclined to override that
with a custom value in consumer of the data anyway.